refactor(ui): rename test files of SubmitButton component
