+++ packages/ui/src/Buttons/LoadingButton/__test__/SubmitButton.snapshot.test.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import renderer, { create } from "react-test-renderer";
 import { expect, test } from "vitest";
 
-import LoadingButton from "../../LoadingButton";
+import LoadingButton from "..";
 
 function toJson(component: renderer.ReactTestRenderer) {
   const result = component.toJSON();
