+++ packages/ui/src/Buttons/LoadingButton/__test__/SubmitButton.test.tsx
@@ -2,7 +2,7 @@ import { render, screen, waitFor } from "@testing-library/react";
 import React from "react";
 import { expect, test } from "vitest";
 
-import LoadingButton from "../";
+import LoadingButton from "..";
 
 const values = {
   label: "Submit",
